Changes in the announcement

+++ controller/announcement.js
@@ -1,7 +1,12 @@
 const Announcement=require("../models/Announcement");
+const Token=require("../models/Token");
 
 async function getAnnouncementAll(req, res) {
   try {
+     const token = await Token.findOne({ encrypted: req.query.adminemailId });
+     if (!token) {
+       return res.status(400).send("Not yet logged in");
+     }
     const announcement = await Announcement.find();
     res.status(200).json(announcement);
   } catch (error) {
@@ -9,23 +14,41 @@ async function getAnnouncementAll(req, res) {
     res.status(500).json({ error: "Internal server error" });
   }
 }
-
 const addAnnouncement = async (req, res) => {
-  const { headline, description } = req.body;
-
   try {
+    // Check if the admin is logged in
+    const token = await Token.findOne({ encrypted: req.query.adminemailId });
+    if (!token) {
+      return res.status(400).send("Not yet logged in");
+    }
+
+    // Extract headline and description from the request body
+    const { headline, description } = req.body;
+
+    // Create a new Announcement instance
     const newAnnouncement = new Announcement({
-      headline,description
+      headline,
+      description,
     });
+
+    // Save the new announcement to the database
     const announcement = await newAnnouncement.save();
+
+    // Send a success response
+    res.status(200).json(announcement);
   } catch (err) {
     console.error(err.message);
     res.status(500).send("Server Error");
   }
 };
 
+
 async function deleteAnnouncement(req, res) {
   try {
+     const token = await Token.findOne({ encrypted: req.query.adminemailId });
+     if (!token) {
+       return res.status(400).send("Not yet logged in");
+     }
     const headline = req.query.headline;
     const deletedAnnouncement = await Announcement.findOneAndDelete({
       headline: headline,
